Add a game end check.
Initially tried doing a template match for the play button to see if the
game had ended, which worked but ran at 5 fps.
Switched to just taking the mean of the gamefield as the gamefield gets
darker when the game ends.
Am now using mean of header as it is smaller and amounts to ~5 fps
improvement.